refactor: 불필요한 decorate 제거
