+++ packages/server/src/admin/admin.controller.ts
@@ -1,12 +1,8 @@
 import { Body, Controller, Get, Post } from "@nestjs/common";
 import { AdminCredential } from "src/auth/dto/adminCredential.dto";
-import { Authority } from "src/commons/constants/enums";
-import { Authorities } from "src/commons/decorators/Authorities.decorator";
-import { PageQuery } from "src/commons/decorators/PageQuery.decorator";
-import { UserSession } from "src/commons/decorators/UserSession.decorator";
 import { Admin } from "src/commons/entities/admin.entity";
-import { FindManyOptions } from "typeorm";
 
+import { UserSession } from "../commons/decorators/user-session.decorator";
 import { AdminService } from "./admin.service";
 import { AdminCreateDto } from "./dto/adminCreate.dto";
 
@@ -20,9 +16,8 @@ export class AdminController {
   }
 
   @Get()
-  @Authorities(Authority.Super)
-  async find(@PageQuery() query: FindManyOptions<Admin>): Promise<Admin[]> {
-    return this.adminService.find(query);
+  async find(): Promise<Admin[]> {
+    return this.adminService.find();
   }
 
   @Post()
