--- packages/server/src/commons/decorators/public.decorator.ts
@@ -1,6 +1,0 @@
-import { SetMetadata } from "@nestjs/common";
-
-export const IS_PUBLIC_KEY = "isPublic";
-export const Public = () => {
-  return SetMetadata(IS_PUBLIC_KEY, true);
-};
